updates to buie sunshape model
the Buie model requires a correction to the chi factor based on polynomials implemented elsewhere (borrowed from Tonatiuh). the correction should be calculated once, so I've added the Ambient class back as an object member of SolarField. It was previously just a collection of static methods. The updated model agrees better with other raytrace models.